+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 1.0.5 - 2017-03-10
+
+- Update code in listener and added collapse.
+- Update code in main controller.
+- Update code in admin_controller.
+- Added permission a_dm_qc_manage in listener and language file.
+- Update dm_qc_index_body.html with collapse code.
+
 ### 1.0.4 - 2017-03-10
 
 - Update code in listener.